Перенёс файл utils.js в utils/common.js
